requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.12
Django==3.1.4
django-crispy-forms==1.10.0
django-datatables-too==0.1.3
django-environ==0.4.5
django-extra-views==0.13.0
django-filter==2.4.0
django-multi-form-view==2.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
google-api-core==1.26.3
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
numpy==1.22.3
oauthlib==3.1.0
packaging==20.9
Pillow==9.1.0
protobuf==3.15.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
uritemplate==3.0.1
urllib3==1.26.2
